Provide access to remote invocation target instance
The invocation target is serialized and then send to a separate unblocker worker process. Therefor the unblocker worker process operates on an independent clone of the invocation target. If the invocation target is mutable, any changes will not be reflected by the client's instance.

This change provides access to the state of the invocation target after the task was executed by the unblocker worker.